+++ .devcontainer/devcontainer.json
@@ -1,8 +1,9 @@
 // See https://aka.ms/vscode-remote/devcontainer.json for format details.
 {
-	// "context": "..",
+	"context": "..",
 	"name": "SMHI Dev",
-	"dockerFile": "Dockerfile",
+	"image": "ludeeus/container:python-base",
+	"postCreateCommand": "python3 -m pip install -r requirements.txt",
 	"runArgs": [
 		"-e",
 		"GIT_EDTIOR='code --wait'"
@@ -12,6 +13,7 @@
 		"tabnine.tabnine-vscode"
 	],
 	"settings": {
+		"terminal.integrated.shell.linux": "/bin/bash",
 		"python.pythonPath": "/usr/local/bin/python",
 		"python.linting.pylintEnabled": true,
 		"python.linting.enabled": true,
