+++ .github/workflows/publish.yml
@@ -27,6 +27,5 @@ jobs:
         TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
         TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
       run: |
-        cd smhi || exit 1
         python setup.py sdist bdist_wheel
         twine upload dist/*
