+++ smhi/__init__.py
@@ -1,0 +1,8 @@
+"""intit.py"""
+
+from smhi.smhi_lib import Smhi, SmhiForecast, SmhiAPIBase
+
+__title__ = "SMHI"
+__version__ = "1.0.13"
+__author__ = "helto4real"
+__license__ = "MIT"
